Allow managers to download tickets
